% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countyfips.R
\docType{data}
\name{countyfips}
\alias{countyfips}
\title{County FIPS data}
\format{
Data frame with columns
\describe{
\item{fips, county_name, county_fips, county_ns}{five-digit FIPS, county name, and county GNIS code}
\item{state_name, state_fips, state_abb}{State name, two-digit FIPS code, and postal abbreviation}
}
}
\source{
United States Census Bureau,
\url{https://www.census.gov/library/reference/code-lists/ansi.html}
}
\usage{
countyfips
}
\description{
County FIPS data
}
\examples{
countyfips
}
\keyword{datasets}
